docker/requirements.txt
# GCP dependencies
google-cloud-storage>=2.7.0
google-cloud-pubsub>=2.13.0
google-cloud-firestore>=2.10.0
google-cloud-logging>=3.7.0

# File processing dependencies
ffmpeg-python>=0.2.0
opencv-python>=4.6.0
numpy>=1.23.0

# Compression and archive tools
rosbags>=0.9.14
mcap>=0.0.10
h5py>=3.7.0

# Web and API dependencies
flask>=2.3.0
webargs>=8.2.0
requests>=2.28.0
boto3>=1.26.0

# Utility libraries
PyYAML>=6.0
python-json-logger>=2.0.4
tqdm>=4.64.0
python-dotenv>=1.0.0 